validation

+++ src/main/java/com/alkemy/ong/domain/usecase/impl/CommentServiceImpl.java
@@ -4,16 +4,20 @@ import com.alkemy.ong.common.exception.NotFoundException;
 import com.alkemy.ong.domain.model.Comment;
 import com.alkemy.ong.domain.model.CommentList;
 import com.alkemy.ong.domain.model.New;
+import com.alkemy.ong.domain.model.User;
 import com.alkemy.ong.domain.repository.CommentRepository;
 import com.alkemy.ong.domain.repository.NewRepository;
 import com.alkemy.ong.domain.usecase.CommentService;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -32,11 +36,19 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    @SneakyThrows
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(Long id, User user) {
         Optional<Comment> optional = commentRepository.findById(id);
         if (optional.isPresent()) {
             Comment comment = optional.get();
+
+            boolean canDelete = Objects.equals(user, comment.getUser()) ||
+                    Objects.equals(user.getRole().getAuthority(), "ROLE_ADMIN");
+
+            if (!canDelete) {
+                throw new AccessDeniedException("User not authorized to delete this resource");
+            }
             commentRepository.delete(comment);
         }
     }
